Desktop-App/src/backend/requirements.txt
flask
flask-cors
werkzeug
numpy
scikit-image
pillow